Update chatbotOPHELIA.py
Added code to show OPHELIA's separate counts of each type of emotion word in the dictionary, output to the 'OPHELIAdata.txt' file. Also fixed a small mistake that had slightly corrupted the emotion dictionary file, it was a simple matter of ignoring empty '' words when determining the emotion counts of a message.